Add ship explosion particle effects
